# ndk_frontend/openapi.yaml
openapi: 3.0.0
info:
  title: NUST LMS API
  version: 1.0.0
  description: |+
    NUST Learning Management System API for developer integrations (Moodle AJAX wrapper).
    Authenticate with your NUST LMS credentials via `POST /auth/login` to receive a bearer token, then call any operation with `Authorization: Bearer <token>`. The NustDevKit gateway logs into the LMS on your behalf and manages the underlying session (cookie + sesskey) server-side, so you never handle the ephemeral sesskey.

    ## Getting Started with the SDKs

    1. **Get the SDK** for your language — download the zip, unzip it, and install it:
       - **TypeScript** — [download](./sdks/nust-lms-api-ts_generic_lib.zip), then run `npm install` in the unzipped folder, then add it to your project.
       - **Python** — [download](./sdks/nust-lms-api-python_generic_lib.zip), then run `pip install -r requirements.txt` in the unzipped folder.
       - **Java** — [download](./sdks/nust-lms-api-java_eclipse_jre_lib.zip), then build the JAR with Gradle and add it to your project.
       - **.NET (C#)** — [download](./sdks/nust-lms-api-cs_net_standard_lib.zip), then add a project reference to the `NustLmsApi.Standard` project.
       - **Go** — [download](./sdks/nust-lms-api-go_generic_lib.zip), then add a local `replace` for the folder in your `go.mod`, then run `go get`.
       - **PHP** — [download](./sdks/nust-lms-api-php_generic_lib_v2.zip), then run `composer install` in the unzipped folder.
...

...

  contact:
...
servers:
...
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Gateway-issued JWT. Obtain it from `POST /auth/login` using your NUST LMS credentials, then send it as `Authorization: Bearer <token>`. The gateway manages the underlying LMS session (cookie + sesskey) server-side, so the ephemeral sesskey never leaves the gateway.
  schemas:
...
    CourseTimelineClassification:
      type: string
      description: Timeline classification filter for enrolled courses.
      enum:
        - all
        - inprogress
        - past
        - future
        - favourites
        - hidden
        - allincludinghidden
    CourseTimelineSortField:
      type: string
      description: Field used to sort enrolled course results.
      enum:
        - fullname
        - shortname
        - id
        - idnumber
        - timeaccess
...
    SiteNotification:
      type: object
      description: A site-level notification (structure varies by type).
      properties:
        id:
          type: integer
          description: Unique identifier for the site notification.
        message:
          type: string
          description: Notification message content.
        type:
          type: string
          description: Notification type identifier.
    MoodleError:
      x-scalar-ignore: true
      type: object
      description: Error returned by the LMS on a failed operation.
      properties:
        errorcode:
          type: string
          description: Machine-readable error code identifier.
          example: invalidrecord
        message:
          type: string
          description: Human-readable error description.
          example: Can't find data record in database table.
        link:
          type: string
          description: Redirect URL for the user; empty string if not applicable.
          example: ""
        moreinfourl:
          type: string
          description: URL to additional error documentation; empty string if not available.
          example: ""
        debuginfo:
          type: string
          description: Additional debug information; empty string when not in debug mode.
          example: ""
...
    RecentCourseFields:
      type: object
      description: Additional fields specific to recently accessed courses.
      required:
        - timeaccess
      properties:
        timeaccess:
          type: integer
          description: Unix timestamp of the user's most recent access to this course.
          example: 1781632422
    RecentCourse:
      description: A recently accessed course, extending Course with the last-access timestamp.
      allOf:
        - $ref: "#/components/schemas/Course"
        - $ref: "#/components/schemas/RecentCourseFields"
    EnrolledCoursesResponse:
      type: object
      description: Paginated enrolled-courses response.
      required:
        - courses
        - nextoffset
      properties:
        courses:
          type: array
          description: List of enrolled courses for the current page.
          items:
            $ref: "#/components/schemas/Course"
        nextoffset:
          type: integer
          description: Offset to pass on the next page request; -1 when no further pages exist.
          example: 29
...
    CalendarEventsResponse:
      type: object
      description: A list of calendar events with pagination boundary IDs.
      required:
        - events
        - firstid
        - lastid
      properties:
        events:
          type: array
          description: List of calendar action events for the current result set.
          items:
            $ref: "#/components/schemas/CalendarEvent"
        firstid:
          type: integer
          description: Event ID of the first item in the result set.
          example: 149885
        lastid:
          type: integer
          description: Event ID of the last item in the result set.
          example: 151886
...
paths:
...
  /service/core_course_get_recent_courses:
    get:
      operationId: getCoreRecentCourses
      summary: Get recently accessed courses
      description: |
        Returns the courses the authenticated user has accessed most recently, ordered by last access time.
        Moodle method: `core_course_get_recent_courses`
      tags:
        - Courses
      parameters:
        - name: limit
          in: query
          required: true
          description: Maximum number of courses to return.
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of recently accessed courses ordered by last access.
          content:
            application/json:
              schema:
                type: array
                description: Recently accessed courses ordered by last access time.
                items:
                  $ref: "#/components/schemas/RecentCourse"
        "400":
          description: Bad request or Moodle exception.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodleError"
      x-codeSamples:
        - lang: typescript
          label: TypeScript
          source: |-
            const limit = 10;

            try {
              const response = await coursesApi.getCoreRecentCourses(limit);

              // Extracting fully parsed response body.
              console.log(response.result);

              // Extracting response status code.
              console.log(response.statusCode);
              // Extracting response headers.
              console.log(response.headers);
              // Extracting response body of type `string | Stream`
              console.log(response.body);
            } catch (error) {
              if (error instanceof ApiError) {
                // Extracting response error status code.
                console.log(error.statusCode);
                // Extracting response error headers.
                console.log(error.headers);
                // Extracting response error body of type `string | Stream`.
                console.log(error.body);
                if (error instanceof MoodleError) {
                  console.log(error.result);
                }
              }
            }
        - lang: python
          label: Python
          source: |-
            limit = 10

            result = courses_api.get_core_recent_courses(limit)

            if result.is_success():
                print(result.body)
            elif result.is_error():
                print(result.errors)
        - lang: java
          label: Java
          source: |-
            int limit = 10;

            coursesApi.getCoreRecentCoursesAsync(limit).thenAccept(result -> {
                // TODO success callback handler
                System.out.println(result);
            }).exceptionally(exception -> {
                Throwable cause = exception.getCause();

                if (cause instanceof MoodleErrorException) {
                    MoodleErrorException moodleErrorException = (MoodleErrorException) cause;
                    moodleErrorException.printStackTrace();
                } else {
                    // fallback for unexpected errors
                    exception.printStackTrace();
                }

                return null;
            });
        - lang: csharp
          label: .NET (C#)
          source: |-
            int limit = 10;
            try
            {
                ApiResponse<List<RecentCourse>> result = await coursesApi.GetCoreRecentCoursesAsync(limit);
            }
            catch (ApiException e)
            {
                Console.WriteLine(e.Message);
                if (e is MoodleErrorException)
                {
                   // TODO: Handle MoodleErrorException exception here
                }
            }
        - lang: go
          label: Go
          source: |-
            ctx := context.Background()

            limit := 10

            apiResponse, err := coursesApi.GetCoreRecentCourses(ctx, limit)
            if err != nil {
                switch typedErr := err.(type) {
                    case *errors.MoodleError:
                        log.Fatalln("MoodleErrorException: ", typedErr)
                    default:
                        log.Fatalln(err)
                }
            } else {
                // Printing the result and response
                fmt.Println(apiResponse.Data)
                fmt.Println(apiResponse.Response.StatusCode)
            }
        - lang: php
          label: PHP
          source: |-
            $limit = 10;

            $coursesApi = $client->getCoursesApi();
            $apiResponse = $coursesApi->getCoreRecentCourses($limit);

            // Extracting response status code
            var_dump($apiResponse->getStatusCode());
            // Extracting response headers
            var_dump($apiResponse->getHeaders());

            if ($apiResponse->isSuccess()) {
                echo 'RecentCourse[]:';
                var_dump($apiResponse->getResult());
            } else {
                $error = $apiResponse->getResult();
                var_dump($error);
            }
        - lang: ruby
          label: Ruby
          source: |-
            limit = 10

            result = courses_api.get_core_recent_courses(limit)

            if result.success?
              puts result.data
            elsif result.error?
              warn result.errors
            end
  /service/core_course_get_enrolled_courses_by_timeline_classification:
    get:
      operationId: getEnrolledCoursesByTimeline
      summary: Get enrolled courses by timeline classification
      description: |
        Returns all courses the user is enrolled in, filtered by timeline classification and sorted by a chosen field. Supports pagination via offset.
        Moodle method: `core_course_get_enrolled_courses_by_timeline_classification`
      tags:
        - Courses
      parameters:
        - name: offset
          in: query
          required: true
          description: Zero-based pagination offset.
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: true
          description: Maximum number of courses to return.
          schema:
            type: integer
            default: 50
        - name: classification
          in: query
          required: true
          description: Timeline classification filter for enrolled courses.
          schema:
            $ref: "#/components/schemas/CourseTimelineClassification"
        - name: sort
          in: query
          required: true
          description: Field used to sort enrolled course results.
          schema:
            $ref: "#/components/schemas/CourseTimelineSortField"
      responses:
        "200":
          description: Paginated list of enrolled courses.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrolledCoursesResponse"
        "400":
          description: Bad request or Moodle exception.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodleError"
      x-codeSamples:
        - lang: typescript
          label: TypeScript
          source: |-
            const offset = 0;

            const limit = 50;

            const classification = CourseTimelineClassification.Inprogress;

            const sort = CourseTimelineSortField.Idnumber;

            try {
              const response = await coursesApi.getEnrolledCoursesByTimeline(
                offset,
                limit,
                classification,
                sort
              );

              // Extracting fully parsed response body.
              console.log(response.result);

              // Extracting response status code.
              console.log(response.statusCode);
              // Extracting response headers.
              console.log(response.headers);
              // Extracting response body of type `string | Stream`
              console.log(response.body);
            } catch (error) {
              if (error instanceof ApiError) {
                // Extracting response error status code.
                console.log(error.statusCode);
                // Extracting response error headers.
                console.log(error.headers);
                // Extracting response error body of type `string | Stream`.
                console.log(error.body);
                if (error instanceof MoodleError) {
                  console.log(error.result);
                }
              }
            }
        - lang: python
          label: Python
          source: |-
            offset = 0

            limit = 50

            classification = CourseTimelineClassification.INPROGRESS

            sort = CourseTimelineSortField.IDNUMBER

            result = courses_api.get_enrolled_courses_by_timeline(
                offset,
                limit,
                classification,
                sort
            )

            if result.is_success():
                print(result.body)
            elif result.is_error():
                print(result.errors)
        - lang: java
          label: Java
          source: |-
            int offset = 0;
            int limit = 50;
            CourseTimelineClassification classification = CourseTimelineClassification.INPROGRESS;
            CourseTimelineSortField sort = CourseTimelineSortField.IDNUMBER;

            coursesApi.getEnrolledCoursesByTimelineAsync(offset, limit, classification, sort).thenAccept(result -> {
                // TODO success callback handler
                System.out.println(result);
            }).exceptionally(exception -> {
                Throwable cause = exception.getCause();

                if (cause instanceof MoodleErrorException) {
                    MoodleErrorException moodleErrorException = (MoodleErrorException) cause;
                    moodleErrorException.printStackTrace();
                } else {
                    // fallback for unexpected errors
                    exception.printStackTrace();
                }

                return null;
            });
        - lang: csharp
          label: .NET (C#)
          source: |-
            int offset = 0;
            int limit = 50;
            CourseTimelineClassification classification = CourseTimelineClassification.Inprogress;
            CourseTimelineSortField sort = CourseTimelineSortField.Idnumber;
            try
            {
                ApiResponse<EnrolledCoursesResponse> result = await coursesApi.GetEnrolledCoursesByTimelineAsync(
                    offset,
                    limit,
                    classification,
                    sort
                );
            }
            catch (ApiException e)
            {
                Console.WriteLine(e.Message);
                if (e is MoodleErrorException)
                {
                   // TODO: Handle MoodleErrorException exception here
                }
            }
        - lang: go
          label: Go
          source: |-
            ctx := context.Background()

            offset := 0

            limit := 50

            classification := models.CourseTimelineClassification_Inprogress

            sort := models.CourseTimelineSortField_Idnumber

            apiResponse, err := coursesApi.GetEnrolledCoursesByTimeline(ctx, offset, limit, classification, sort)
            if err != nil {
                switch typedErr := err.(type) {
                    case *errors.MoodleError:
                        log.Fatalln("MoodleErrorException: ", typedErr)
                    default:
                        log.Fatalln(err)
                }
            } else {
                // Printing the result and response
                fmt.Println(apiResponse.Data)
                fmt.Println(apiResponse.Response.StatusCode)
            }
        - lang: php
          label: PHP
          source: |-
            $offset = 0;

            $limit = 50;

            $classification = CourseTimelineClassification::INPROGRESS;

            $sort = CourseTimelineSortField::IDNUMBER;

            $coursesApi = $client->getCoursesApi();
            $apiResponse = $coursesApi->getEnrolledCoursesByTimeline(
                $offset,
                $limit,
                $classification,
                $sort
            );

            // Extracting response status code
            var_dump($apiResponse->getStatusCode());
            // Extracting response headers
            var_dump($apiResponse->getHeaders());

            if ($apiResponse->isSuccess()) {
                echo 'EnrolledCoursesResponse:';
                var_dump($apiResponse->getResult());
            } else {
                $error = $apiResponse->getResult();
                var_dump($error);
            }
        - lang: ruby
          label: Ruby
          source: |-
            offset = 0

            limit = 50

            classification = CourseTimelineClassification::INPROGRESS

            sort = CourseTimelineSortField::IDNUMBER

            result = courses_api.get_enrolled_courses_by_timeline(
              offset,
              limit,
              classification,
              sort
            )

            if result.success?
              puts result.data
            elsif result.error?
              warn result.errors
            end
  /service/core_fetch_notifications:
    get:
      operationId: fetchNotifications
      summary: Fetch site-level notifications for a context
      description: |
        Returns pending site-level notifications for the given Moodle context. Returns an empty array when there are no pending notifications.
        Moodle method: `core_fetch_notifications`
      tags:
        - Notifications
      parameters:
        - name: contextid
          in: query
          required: true
          description: |
            Moodle context ID for which to fetch notifications. The user context ID can be found in the user profile image URL path segment.
          schema:
            type: integer
      responses:
        "200":
          description: Array of site notifications (empty array when none are pending).
          content:
            application/json:
              schema:
                type: array
                description: Site-level notifications for the given context; empty array when none are pending.
                items:
                  $ref: "#/components/schemas/SiteNotification"
        "400":
          description: Bad request or Moodle exception.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodleError"
      x-codeSamples:
        - lang: typescript
          label: TypeScript
          source: |-
            const contextid = 156;

            try {
              const response = await notificationsApi.fetchNotifications(contextid);

              // Extracting fully parsed response body.
              console.log(response.result);

              // Extracting response status code.
              console.log(response.statusCode);
              // Extracting response headers.
              console.log(response.headers);
              // Extracting response body of type `string | Stream`
              console.log(response.body);
            } catch (error) {
              if (error instanceof ApiError) {
                // Extracting response error status code.
                console.log(error.statusCode);
                // Extracting response error headers.
                console.log(error.headers);
                // Extracting response error body of type `string | Stream`.
                console.log(error.body);
                if (error instanceof MoodleError) {
                  console.log(error.result);
                }
              }
            }
        - lang: python
          label: Python
          source: |-
            contextid = 156

            result = notifications_api.fetch_notifications(contextid)

            if result.is_success():
                print(result.body)
            elif result.is_error():
                print(result.errors)
        - lang: java
          label: Java
          source: |-
            int contextid = 156;

            notificationsApi.fetchNotificationsAsync(contextid).thenAccept(result -> {
                // TODO success callback handler
                System.out.println(result);
            }).exceptionally(exception -> {
                Throwable cause = exception.getCause();

                if (cause instanceof MoodleErrorException) {
                    MoodleErrorException moodleErrorException = (MoodleErrorException) cause;
                    moodleErrorException.printStackTrace();
                } else {
                    // fallback for unexpected errors
                    exception.printStackTrace();
                }

                return null;
            });
        - lang: csharp
          label: .NET (C#)
          source: |-
            int contextid = 156;
            try
            {
                ApiResponse<List<SiteNotification>> result = await notificationsApi.FetchNotificationsAsync(contextid);
            }
            catch (ApiException e)
            {
                Console.WriteLine(e.Message);
                if (e is MoodleErrorException)
                {
                   // TODO: Handle MoodleErrorException exception here
                }
            }
        - lang: go
          label: Go
          source: |-
            ctx := context.Background()

            contextid := 156

            apiResponse, err := notificationsApi.FetchNotifications(ctx, contextid)
            if err != nil {
                switch typedErr := err.(type) {
                    case *errors.MoodleError:
                        log.Fatalln("MoodleErrorException: ", typedErr)
                    default:
                        log.Fatalln(err)
                }
            } else {
                // Printing the result and response
                fmt.Println(apiResponse.Data)
                fmt.Println(apiResponse.Response.StatusCode)
            }
        - lang: php
          label: PHP
          source: |-
            $contextid = 156;

            $notificationsApi = $client->getNotificationsApi();
            $apiResponse = $notificationsApi->fetchNotifications($contextid);

            // Extracting response status code
            var_dump($apiResponse->getStatusCode());
            // Extracting response headers
            var_dump($apiResponse->getHeaders());

            if ($apiResponse->isSuccess()) {
                echo 'SiteNotification[]:';
                var_dump($apiResponse->getResult());
            } else {
                $error = $apiResponse->getResult();
                var_dump($error);
            }
        - lang: ruby
          label: Ruby
          source: |-
            contextid = 156

            result = notifications_api.fetch_notifications(contextid)

            if result.success?
              puts result.data
            elsif result.error?
              warn result.errors
            end
  /service/core_calendar_get_action_events_by_timesort:
    get:
      operationId: getCalendarEventsByTimesort
      summary: Get global calendar action events sorted by time
      description: |
        Returns action events (deadlines) across all enrolled courses, ordered by their sort timestamp. On NUST LMS these are submission due dates for assignments, labs, quizzes, and similar dated activities. Use `timesortfrom` to filter to upcoming deadlines only.
        Moodle method: `core_calendar_get_action_events_by_timesort`
      tags:
        - Calendar
      parameters:
        - name: limitnum
          in: query
          required: true
          description: Maximum number of events to return.
          schema:
            type: integer
        - name: timesortfrom
          in: query
          required: true
          description: Only return events with timesort >= this Unix timestamp. Pass 0 for no lower bound.
          schema:
            type: integer
        - name: timesortto
          in: query
          required: false
          description: Only return events with timesort <= this Unix timestamp.
          schema:
            type: integer
        - name: aftereventid
          in: query
          required: false
          description: Return events whose ID is greater than this value (cursor pagination).
          schema:
            type: integer
      responses:
        "200":
          description: Calendar events sorted by timesort.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEventsResponse"
        "400":
          description: Bad request or Moodle exception.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodleError"
      x-codeSamples:
        - lang: typescript
          label: TypeScript
          source: |-
            const limitnum = 32;

            const timesortfrom = 58;

            try {
              const response = await calendarApi.getCalendarEventsByTimesort(
                limitnum,
                timesortfrom
              );

              // Extracting fully parsed response body.
              console.log(response.result);

              // Extracting response status code.
              console.log(response.statusCode);
              // Extracting response headers.
              console.log(response.headers);
              // Extracting response body of type `string | Stream`
              console.log(response.body);
            } catch (error) {
              if (error instanceof ApiError) {
                // Extracting response error status code.
                console.log(error.statusCode);
                // Extracting response error headers.
                console.log(error.headers);
                // Extracting response error body of type `string | Stream`.
                console.log(error.body);
                if (error instanceof MoodleError) {
                  console.log(error.result);
                }
              }
            }
        - lang: python
          label: Python
          source: |-
            limitnum = 32

            timesortfrom = 58

            result = calendar_api.get_calendar_events_by_timesort(
                limitnum,
                timesortfrom
            )

            if result.is_success():
                print(result.body)
            elif result.is_error():
                print(result.errors)
        - lang: java
          label: Java
          source: |-
            int limitnum = 32;
            int timesortfrom = 58;

            calendarApi.getCalendarEventsByTimesortAsync(limitnum, timesortfrom, null, null).thenAccept(result -> {
                // TODO success callback handler
                System.out.println(result);
            }).exceptionally(exception -> {
                Throwable cause = exception.getCause();

                if (cause instanceof MoodleErrorException) {
                    MoodleErrorException moodleErrorException = (MoodleErrorException) cause;
                    moodleErrorException.printStackTrace();
                } else {
                    // fallback for unexpected errors
                    exception.printStackTrace();
                }

                return null;
            });
        - lang: csharp
          label: .NET (C#)
          source: |-
            int limitnum = 32;
            int timesortfrom = 58;
            try
            {
                ApiResponse<CalendarEventsResponse> result = await calendarApi.GetCalendarEventsByTimesortAsync(
                    limitnum,
                    timesortfrom
                );
            }
            catch (ApiException e)
            {
                Console.WriteLine(e.Message);
                if (e is MoodleErrorException)
                {
                   // TODO: Handle MoodleErrorException exception here
                }
            }
        - lang: go
          label: Go
          source: |-
            ctx := context.Background()

            limitnum := 32

            timesortfrom := 58

            apiResponse, err := calendarApi.GetCalendarEventsByTimesort(ctx, limitnum, timesortfrom, nil, nil)
            if err != nil {
                switch typedErr := err.(type) {
                    case *errors.MoodleError:
                        log.Fatalln("MoodleErrorException: ", typedErr)
                    default:
                        log.Fatalln(err)
                }
            } else {
                // Printing the result and response
                fmt.Println(apiResponse.Data)
                fmt.Println(apiResponse.Response.StatusCode)
            }
        - lang: php
          label: PHP
          source: |-
            $limitnum = 32;

            $timesortfrom = 58;

            $calendarApi = $client->getCalendarApi();
            $apiResponse = $calendarApi->getCalendarEventsByTimesort(
                $limitnum,
                $timesortfrom
            );

            // Extracting response status code
            var_dump($apiResponse->getStatusCode());
            // Extracting response headers
            var_dump($apiResponse->getHeaders());

            if ($apiResponse->isSuccess()) {
                echo 'CalendarEventsResponse:';
                var_dump($apiResponse->getResult());
            } else {
                $error = $apiResponse->getResult();
                var_dump($error);
            }
        - lang: ruby
          label: Ruby
          source: |-
            limitnum = 32

            timesortfrom = 58

            result = calendar_api.get_calendar_events_by_timesort(
              limitnum,
              timesortfrom
            )

            if result.success?
              puts result.data
            elsif result.error?
              warn result.errors
            end
...